separate sign tx into two different components
